Implement logic for finding most/least common binary
Techinally 'wip' for part 2. I left a TODO comment in `part2` for where
to pick up.

Lots of code and trial & error went into this, and I know there is room
for drying up the most/least recursive methods. Right now I'm out of
time and brainpower to continue, though.